member can input answer now

+++ components/poll-detail/DeletePollDetail.php
@@ -1,7 +1,12 @@
 <?php
 $export = function () {
-    $db = new Database;
-    $poll_detail_id = $_POST['delete_poll_post'];
-    $db->deletePollDetail_ByID($poll_detail_id);
-    return true;
+    if (!empty($_POST['delete_poll_post'])) {
+
+        $db = new Database;
+        $poll_detail_id = $_POST['delete_poll_post'];
+        $db->deletePollDetail_ByID($poll_detail_id);
+        return true;
+    } else {
+        return false;
+    }
 };
